Move server initialization to its own file
