+++ lib/init-electron.js
@@ -22,6 +22,7 @@ in an electron context
     return
   }
 
+  logger.debug('Initializing window')
   electron.initWindow(`http://localhost:${UserDefaults.data.httpPort}`)
 })()
